Move TableBase from models/database/base.py to models/base.py and update imports
