+++ src/local_newsifier/models/__init__.py
@@ -4,13 +4,13 @@
 from sqlmodel import SQLModel
 
 # Export table base
-from local_newsifier.models.database.base import TableBase
+from local_newsifier.models.base import TableBase
 
 # Import all models from their original locations but don't re-export
 # This prevents duplicate class registrations
-from local_newsifier.models.database.article import Article
-from local_newsifier.models.database.entity import Entity
-from local_newsifier.models.database.analysis_result import AnalysisResult
+from local_newsifier.models.article import Article
+from local_newsifier.models.entity import Entity
+from local_newsifier.models.analysis_result import AnalysisResult
 from local_newsifier.models.entity_tracking import (
     CanonicalEntity,
     EntityMention,
